+++ buildpacks/ruby/src/gem_list.rs
@@ -1,10 +1,10 @@
-use commons::fun_run::{CmdError, CommandWithName};
 use commons::gem_version::GemVersion;
 use commons::output::{
     fmt,
     section_log::{log_step_timed, SectionLogger},
 };
 use core::str::FromStr;
+use fun_run::{CmdError, CommandWithName};
 use regex::Regex;
 use std::collections::HashMap;
 use std::ffi::OsStr;