+++ buildpacks/ruby/src/layers/bundle_install_layer.rs
@@ -4,13 +4,11 @@ use commons::output::{
 };
 
 use crate::{BundleWithout, RubyBuildpack, RubyBuildpackError};
-use commons::fun_run::CommandWithName;
 use commons::{
-    display::SentenceList,
-    fun_run::{self, CmdError},
-    gemfile_lock::ResolvedRubyVersion,
-    metadata_digest::MetadataDigest,
+    display::SentenceList, gemfile_lock::ResolvedRubyVersion, metadata_digest::MetadataDigest,
 };
+use fun_run::CommandWithName;
+use fun_run::{self, CmdError};
 use libcnb::{
     build::BuildContext,
     data::{buildpack::StackId, layer_content_metadata::LayerTypes},