+++ buildpacks/ruby/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- The `fun_run` commons library was moved to it's own crate ()
+
 ## [2.1.2] - 2023-10-31
 
 ### Fixed